requirements.txt
streamlit>=1.32.0
pyyaml>=6.0
Pillow>=10.0
